Allow use a regex to match labels to stickers
We want to add more flexibility to match different labels and
represent as stickers, for instance to cover the case
of "lead x", "team y". We want to allow multiple spaces.

Different labels would render in different values for the
Title, Conment and Image.

We add an additional attribute `regex` for each sticker to be
matched with the card labels, and allow to use regex matching
to populate the values of Title, Image or Comment.

For instance:

   - name: team
     regex: 'team:?\s*([a-zA-Z0-9\s]+)'
     title: "Team $1"

   - name: quarter
     regex: '^2018q([1-4])$'
     image: 'quarter_$1.png'
     title: "Q$1"