Add classes to the stickers HTML
We want to be able to filter cards in javascript based on the
stickers associated.

We classes with the sticker name to the sticker `div` so that
we can use jquery to search the right cards.

In the case of stickers with regex, we allow define an additional
class attribute with regex matchers.