Add buttons to filter teams A, B, C
This is the MVP for the filtering. We hardcoded buttons to filter
based on the different teams: A, B, C or "Show All".

In the future it is desirable that these buttons are generated
dynamically.

The implementation is a simple button with a JS that would use
jQuery to:
 - Hide all the cards with a `sticker-team` class and
not a `sticker-team-${name}` class
 - Show all the cards without `stickert-team` or with
 `sticker-team-${name}`

We use the `gracefulIn()` and `gracefulOut()` to provide nice
transitions.