Process blocker messages and look for a scheduled date
If the blocker message contains any of the following expressions, then a
scheduled sticker will be attached to the story:
 - until 2/9
 - until 02/09
 - until 02/09/2019
 - until 2019/09/02
 - until 2019-09-02

Other prepositions are also supported: "on or after", "after, "before", "on"

If "before" is used then the scheduled day will be the next day.

We will stop at the first successfully parsed message, so any additional dates are ignored.

We will only add a scheduled sticker if the scheduled date is in the future.

The blocker sticker won't be shown if there are no other blockers than one containing a scheduled date.