Visually indicate applied filter
In order to be able to know which filter was applied, a type for the filters
had to be introduced. The applied include and exclude filters were also
recorded on the response, so that they could be used in the template.

The template now excludes the `govuk-button--secondary` class from the filters
buttons IFF the filter is applied. Otherwise, the class is added.