Make applied filters visually distinct
Buttons for filters which were applied appeared identical to those which were
not, making it confusing. This commit makes applied filters visually distinct.

A new `Filter` interface has been introduced, with implementations for each
type of filter. The implementations are responsible for determining if they
have been applied, based on supplied query texts.